requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==4.0.1
django-heroku==0.3.1
django-tinymce==3.4.0
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0
